% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDN_LC_STOCKS_DESC_DATA.R
\docType{data}
\name{CDN_LC_STOCKS_DESC_DATA}
\alias{CDN_LC_STOCKS_DESC_DATA}
\title{CDN_LC_STOCKS_DESC_DATA}
\format{An object of class \code{data.frame} with 50 rows and 5 columns.}
\usage{
data(CDN_LC_STOCKS_DESC_DATA)
}
\description{
Descriptive data (sector, mkt cap, etc) for a random sample of 50
Canadian stocks. Matched with the returns data set.
}
\examples{
data(CDN_LC_STOCKS_DESC_DATA)

}
\keyword{datasets}
