% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDN_LC_STOCKS_RETURN_DATA.R
\docType{data}
\name{CDN_LC_STOCKS_RETURN_DATA}
\alias{CDN_LC_STOCKS_RETURN_DATA}
\title{CDN_LC_STOCKS_RETURN_DATA}
\format{An object of class \code{data.frame} with 793 rows and 51 columns.}
\usage{
data(CDN_LC_STOCKS_RETURN_DATA)
}
\description{
Daily total returns (gross of dividends) for a random sample of 50
Canadian stocks over the period  2015-01-02 to 2018-02-28. May
contain N/A observations.
}
\examples{
data(CDN_LC_STOCKS_RETURN_DATA)

}
\keyword{datasets}
