% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateHedge.R
\name{EstimateHedge}
\alias{EstimateHedge}
\title{EstimateHedge}
\usage{
EstimateHedge(target_var, hedge_vars)
}
\arguments{
\item{target_var}{A data frame of dimension t x 1 containing the target
variable to be hedged.}

\item{hedge_vars}{A data frame of dimension t x k containing k variables
or istruments to be used within the hedge portfolio.}
}
\value{
A list object containing the constrained hedge ratios, the
unconstrained hedge ratios, the r-squared of both the constrained and
unconstrained solutions, as well as the efficiency loss (pct reduction in Rsq)
from the edge constraint.
}
\description{
Calculate optimal hedge ratio coefficients for a set of instruments
with the constraint that the sum of the coefficients equals 1
}
