% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton.R
\name{Newton}
\alias{Newton}
\title{Newton()}
\usage{
Newton(f, h = 0.01, tol = 1e-12, x0 = 1, maxiter)
}
\arguments{
\item{f}{A function to be evaluated.}

\item{h}{Optional parameter. The step size. Default 0.01.}

\item{tol}{Optional parameter. The error tolerance for the function
appriximation. Default 1e-12.}

\item{x0}{Optional parameter. Initial value. Default 1.}

\item{maxiter}{Maximum number of iterations before exit.}
}
\value{
List containing iteration values, the estimated root, and the
number of terations used to estimate the root.
}
\description{
Find the root of a polynomial function.
}
