% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandPorts.R
\name{RandPorts}
\alias{RandPorts}
\title{RandPorts()}
\usage{
RandPorts(num_ports, num_assets)
}
\arguments{
\item{num_ports}{Scalar input number of desired portfolios.}

\item{num_assets}{Scalar input nummber of assets within each portfolio.}
}
\value{
A numeric matrix of dimension num_assets x num_ports containing the
random portfolio weights.
}
\description{
Returns a matrix of num_ports randomly generated portfolio weights that are
non-negative and sum to 1.
}
