% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolveRiskParity.R
\name{SolveRiskParity}
\alias{SolveRiskParity}
\title{SolveRiskParity()}
\usage{
SolveRiskParity(sigma, err_tol = 1e-08)
}
\arguments{
\item{sigma}{A variance-covariance matrix for assets in the investment universe.}

\item{err_tol}{Optional parameter. Tolerance for exiting while loop (may
experience convergence issues with large problems - if results do not
produce expected pctr_i = pctr_j forall i, j then test smaller error
err_tol (default = 0.00000001).}
}
\value{
A list item containing 'wts' (a vector of weights for the resulting
optimal ERC portfolio), 'mctr' (the marginal contributions to risk for each
asset class for the ERC portfolio), 'pctr' (the percent contributions to risk
for each asset class in the ERC portfolio), 'lambda' (optimal lambda).
}
\description{
Finds the risk parity portfolio for a given covariance matrix (defined as
an equal risk contribution portfolio where pctr_i = pctr_j for all i, j).
Uses "Algorithm 1" of Chaves et al (2012) "Efficient Algorithms for
Computing Risk Parity Portfolio Weights".
}
